requirements.txt
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-auth==2.22.0
google-generativeai==0.1.0
googleapis-common-protos==1.60.0
python-dotenv-1.0.0